增加日志文件切换相关API

+++ edap-log/src/test/java/io/edap/log/test/TestLog.java
@@ -6,6 +6,7 @@ import io.edap.log.Logger;
 import io.edap.log.LoggerManager;
 import io.edap.log.config.ConfigManager;
 import io.edap.log.test.spi.EdapTestAdapter;
+import io.edap.util.EdapTime;
 import org.junit.jupiter.api.Test;
 import org.xml.sax.SAXException;
 
@@ -15,8 +16,9 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.util.Iterator;
-import java.util.ServiceLoader;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -28,19 +30,91 @@ public class TestLog {
     @Test
     public void testLog() throws ParserConfigurationException, IOException, SAXException {
         EdapTestAdapter edapTestAdapter = (EdapTestAdapter)ConfigManager.getLogAdapter();
+        File f = new File("./edap.log");
+        if (f.exists()) {
+            f.delete();
+        }
         if (edapTestAdapter != null) {
             edapTestAdapter.reloadConfig("/edap-log-basefile.xml");
         }
 
         LOG.info("name: {},height: {}", l -> l.arg("edap").arg(90.0));
 
-        File f = new File("./edap.log");
+        f = new File("./edap.log");
         String log = readFile(f);
         assertNotNull(log);
-        f.delete();
+        if (f.exists()) {
+            f.delete();
+        }
         assertEquals(log.substring(23), " INFO  [main] [io.edap.log.test.TestLog] [] []  - name: edap,height: 90.0 \n");
 
         System.out.println(EdapLogContext.instance());
+    }
+
+    @Test
+    public void testFileAppender() throws ParserConfigurationException, IOException, SAXException, ParseException {
+        EdapTestAdapter edapTestAdapter = (EdapTestAdapter)ConfigManager.getLogAdapter();
+        File f = new File("./edap-fileappender.log");
+        if (f.exists()) {
+            f.delete();
+        }
+        if (edapTestAdapter != null) {
+            edapTestAdapter.reloadConfig("/edap-log-fileappender.xml");
+        }
+
+        long start = EdapTime.instance().currentTimeMillis();
+        LOG.info("name: {},height: {}", l -> l.arg("edap").arg(90.0));
+        long end = EdapTime.instance().currentTimeMillis();
+        System.out.println("start:" + start);
+        System.out.println("end:" + end);
+        f = new File("./edap-fileappender.log");
+        String log = readFile(f);
+        assertNotNull(log);
+        String date = log.substring(0, 23);
+        Date logDate = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS").parse(date);
+        System.out.println("date:" + date);
+        System.out.println("logDate.getTime():" + logDate.getTime());
+        assertEquals(logDate.getTime() >= start && logDate.getTime() <= end, true);
+        if (f.exists()) {
+            f.delete();
+        }
+        assertEquals(log.substring(23), " INFO  [main] [io.edap.log.test.TestLog] [] []  - name: edap,height: 90.0 \n");
+
+        System.out.println(EdapLogContext.instance());
+
+    }
+
+    @Test
+    public void testFileAppenderPrudent() throws ParserConfigurationException, IOException, SAXException, ParseException {
+        EdapTestAdapter edapTestAdapter = (EdapTestAdapter)ConfigManager.getLogAdapter();
+        File f = new File("./edap-fileappenderprudent.log");
+        if (f.exists()) {
+            f.delete();
+        }
+        if (edapTestAdapter != null) {
+            edapTestAdapter.reloadConfig("/edap-log-fileappenderprudent.xml");
+        }
+
+        long start = EdapTime.instance().currentTimeMillis();
+        LOG.info("name: {},height: {}", l -> l.arg("edap").arg(90.0));
+        long end = EdapTime.instance().currentTimeMillis();
+        System.out.println("start:" + start);
+        System.out.println("end:" + end);
+        f = new File("./edap-fileappenderprudent.log");
+        String log = readFile(f);
+        assertNotNull(log);
+        String date = log.substring(0, 23);
+        Date logDate = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS").parse(date);
+        System.out.println("date:" + date);
+        System.out.println("logDate.getTime():" + logDate.getTime());
+        assertEquals(logDate.getTime() >= start && logDate.getTime() <= end, true);
+        if (f.exists()) {
+            f.delete();
+        }
+        assertEquals(log.substring(23), " INFO  [main] [io.edap.log.test.TestLog] [] []  - name: edap,height: 90.0 \n");
+
+        System.out.println(EdapLogContext.instance());
+
     }
 
     private static String readFile(File f) throws IOException {
